Steering is done
I have noticed that if the car goes too far from the center of the world the suspension wobbles causing the car to shoot in the air. It might be because of the huge mass of 1 ton of the car, the use of flats instead of doubles or the box collider being too big.